Moved error type to its own module.
